requirements.txt
flask
openai
python-dotenv
python-docx
azure-storage-blob
mcp
fastmcp
gunicorn